@@ -15,50 +15,50 @@
     "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "dependencies": {
-    "@angular/animations": "9.0.5",
-    "@angular/common": "9.0.5",
-    "@angular/compiler": "9.0.5",
-    "@angular/core": "9.0.5",
-    "@angular/forms": "9.0.5",
-    "@angular/platform-browser": "9.0.5",
-    "@angular/platform-browser-dynamic": "9.0.5",
-    "@angular/router": "9.0.5",
+    "@angular/animations": "10.0.3",
+    "@angular/common": "10.0.3",
+    "@angular/compiler": "10.0.3",
+    "@angular/core": "10.0.3",
+    "@angular/forms": "10.0.3",
+    "@angular/platform-browser": "10.0.3",
+    "@angular/platform-browser-dynamic": "10.0.3",
+    "@angular/router": "10.0.3",
     "core-js": "^2.4.1",
     "date-fns": "^1.30.1",
-    "ngx-bootstrap": "^5.5.0",
+    "ngx-bootstrap": "^5.6.1",
     "rxjs": "^6.5.4",
     "ts-md5": "^1.2.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.5",
-    "@angular/cli": "9.0.5",
-    "@angular/compiler-cli": "9.0.5",
-    "@angular/language-service": "9.0.5",
-    "@braune-digital/form-utilities": "^1.1.12",
-    "@ng-select/ng-select": "^2.16.3",
+    "@angular-devkit/build-angular": "~0.1000.2",
+    "@angular/cli": "10.0.2",
+    "@angular/compiler-cli": "10.0.3",
+    "@angular/language-service": "10.0.3",
+    "@braune-digital/form-utilities": "^2.0.2",
+    "@ng-select/ng-select": "^2.20.5",
     "@ngx-translate/core": "^11.0.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^9.0.0",
-    "ng2-ckeditor": "^1.2.2",
-    "ngx-quill": "^4.4.2",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^10.0.0",
+    "ng2-ckeditor": "^1.2.7",
+    "ngx-quill": "^4.8.0",
     "ngx-textarea-autosize": "^2.0.0",
-    "protractor": "~5.1.2",
+    "protractor": "~7.0.0",
     "quill": "^1.3.6",
     "rimraf": "^2.6.2",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "^3.7.5"
+    "tslint": "~6.1.0",
+    "typescript": "^3.9.6"
   }
 }
